License project under Mozilla Public License 2.0
It is important to define the license under which a software project is released in order to permit its use.

Previously, most of the code files of this library did not contain any information about licensing. It is now offered
under the formal open source Mozilla Public License 2.0.

The file src/Arduino_SensorKit.h contained an informal statement releasing its contents into the public domain. That
statement and offer remains, but the content is now also offered under MPL 2.0 in order to support its usage in
applications where a formal open source code license is required.

We did our best to assess the intellectual property of this codebase to ensure that all contributors granted copyright
to Arduino SA. For any inquiry, report, or claim please contact us at license@arduino.cc.

The 3rd party libraries "Grove - Barometer Sensor BMP280", "Grove Temperature And Humidity Sensor", and "U8g2" are
vendored in the project. The licensing of those libraries under permissive licenses was already defined in the
individual folder for each library and their licensing remains unchanged.